Updated g++ to test Makefile on windows. (you need at least 4.9.3 to compile this) Changed instances of #pragma once to the old #ifndef preprocessor calls to add some more 'legacy' support to some compilers. Added blank newlines to files so git doesn't complain.

+++ DracheVM/Misc_Utils.cpp
@@ -40,7 +40,7 @@ bool is_string(std::string str)
 bool is_declared(std::map<std::string, int16_t> &_label_map, std::string _label)
 {
 	// iterate through the map checking for the label.
-	for (std::pair<std::string, int> pair : _label_map) 
+	for (std::pair<std::string, int16_t> pair : _label_map) 
 	{
 		if (pair.first == _label)
 			return true;
